MLIBZ-2266 Separate sql module from sqlite persister. Will do the same with WebSql persister in the future.
