+++ src/nativescript/client.ts
@@ -49,4 +49,8 @@ export class Client extends CoreClient {
     client.activeUserStorage = new ActiveUserStorage();
     return client;
   }
+
+  static sharedInstance() {
+    return super.sharedInstance();
+  }
 }
