+++ src/nativescript/offline-data-storage.ts
@@ -2,11 +2,14 @@ import {
   repositoryProvider,
   storageType,
   KeyValueStoreOfflineRepository,
-  SqliteKeyValueStorePersister,
+  SqliteKeyValueStorePersister
 } from '../core/datastore';
+import { Client } from './client';
+import { NativescriptSqliteModule } from './nativescript-sqlite-module';
 
 const sqliteStorageBuilder = (queue) => {
-  const persister = new SqliteKeyValueStorePersister();
+  const sqlModule = new NativescriptSqliteModule(Client.sharedInstance().appKey);
+  const persister = new SqliteKeyValueStorePersister(sqlModule);
   return new KeyValueStoreOfflineRepository(persister, queue);
 };
